+++ src/lib/services/feed/feed_entry.dart
@@ -1,6 +1,7 @@
 import 'package:mediadrip/services/feed/feed_media.dart';
 import 'package:mediadrip/services/feed/helper.dart';
 import 'package:xml/xml.dart';
+import 'package:intl/intl.dart';
 
 class FeedEntry {
   String link;
@@ -17,9 +18,24 @@ class FeedEntry {
 
   factory FeedEntry.parse(XmlElement element) {
     return FeedEntry(
-      link: findElementOrNull(element, 'link')?.text,
-      published: DateTime.parse(findElementOrNull(element, 'published').text),
+      link: findElementOrNull(element, 'link')?.getAttribute('href'),
+      published: DateTime.parse(findElementOrNull(element, 'published')?.text),
       media: FeedMedia.parse(element)
     );
+  }
+
+  /// this will eventually be removed/refactored to a general 
+  /// media object for the rest of the app to use
+  List<String> modelToCollection() {
+    // formats DateTime i.e. June 18, 2020
+    var dateFormat = DateFormat.yMMMMd('en_US').add_jm();
+
+    return [
+      this.link,
+      this.media.title,
+      this.media.description,
+      (this.published != null) ? dateFormat.format(this.published) : '',
+      this.media.thumbnail.url
+    ];
   }
 }