+++ src/lib/views/view.dart
@@ -5,5 +5,7 @@ abstract class View extends StatelessWidget {
   String get routeAddress;
   IconData get icon;
 
+  List<String> routeArguments = List();
+
   Widget build(BuildContext context);
 }